#14 added some basic routing, the chat.js and about.js
