+++ src/App.js
@@ -1,139 +1,21 @@
-import React, { useRef, useState } from 'react';
+import React from 'react';
 import './App.css';
-import firebase from 'firebase/app';
-import 'firebase/auth';
-import 'firebase/firestore';
-// import 'firebase/analytics';
+import { Route, BrowserRouter as Router, Switch } from "react-router-dom";
 
-import { useAuthState } from 'react-firebase-hooks/auth';
-import { useCollectionData } from 'react-firebase-hooks/firestore';
-require ('dotenv').config();
+import Chat from "./pages/Chat";
+import About from "./pages/About";
 
-// initalise the config of the firebase app
-firebase.initializeApp({
-  apiKey: process.env.REACT_APP_apiKey,
-  authDomain: process.env.REACT_APP_authDomain,
-  projectId: process.env.REACT_APP_projectId,
-  storageBucket: process.env.REACT_APP_storageBucket,
-  messagingSenderId: process.env.REACT_APP_messagingSenderId,
-  appId: process.env.REACT_APP_appId,
-  measurementId: process.env.REACT_APP_measurementId
-
-});
-
-// // catch errors for dotenv
-// if (env.error) {
-//   throw env.error
-// };
-
-const auth = firebase.auth();
-const firestore = firebase.firestore();
-// const analytics = firebase.analytics();
-
-
-function App() {
-  const [user] = useAuthState(auth);
-
-  return (
-    <div className="App">
-      <header>
-        <h1>⚛️🔥💬</h1>
-        <SignOut />
-      </header>
-
-      <section>
-        {user ? <ChatRoom /> : <SignIn />}
-      </section>
-
-    </div>
+function App(){
+  return(
+    <Router>
+      <div className="container">
+      </div>
+      <Switch>
+        <Route path="/" exact component={Chat} />
+        <Route path="/about" component={About} />
+      </Switch>
+    </Router>
   );
 };
-
-function SignIn() {
-
-  const signInWithGoogle = () => {
-    const provider = new firebase.auth.GoogleAuthProvider();
-    auth.signInWithPopup(provider);
-  };
-
-  return (
-    <>
-      <button className="sign-in" onClick={signInWithGoogle}>Sign in with Google</button>
-      <p>Do not violate the community guidelines or you will be banned for life!</p>
-    </>
-  )
-
-};
-
-function SignOut() {
-  return auth.currentUser && (
-    <button className="sign-out" onClick={() => auth.signOut()}>Sign Out</button>
-  )
-};
-
-
-function ChatRoom() {
-  const scrolls = useRef();
-  const messagesRef = firestore.collection('messages');
-  const query = messagesRef.orderBy('createdAt').limit(25);
-
-  const [messages] = useCollectionData(query, { idField: 'id' });
-
-  const [formValue, setFormValue] = useState('');
-
-  
-  // e as event for the function argument
-  const sendMessage = async (e) => {
-    // normally when a form submited it refresh the page but here we add this line to prevent it 
-    e.preventDefault();
-
-    const { uid, photoURL } = auth.currentUser;
-
-    await messagesRef.add({
-      text: formValue,
-      createdAt: firebase.firestore.FieldValue.serverTimestamp(),
-      uid,
-      photoURL
-    })
-
-    setFormValue('');
-
-    // to scrolls the page when user types
-    scrolls.current.scrollIntoView({ behavior: 'smooth' });
-  }
-
-  return (<>
-    <main>
-
-      {messages && messages.map(msg => <ChatMessage key={msg.id} message={msg} />)}
-
-      <span ref={scrolls}></span>
-
-    </main>
-
-    <form onSubmit={sendMessage}>
-
-      <input value={formValue} onChange={(e) => setFormValue(e.target.value)} placeholder="say something nice" />
-
-      <button type="submit" disabled={!formValue}>🕊️</button>
-
-    </form>
-  </>)
-};
-
-
-function ChatMessage(props) {
-  const { text, uid, photoURL } = props.message;
-
-  const messageClass = uid === auth.currentUser.uid ? 'sent' : 'received';
-
-  return (<>
-    <div className={`message ${messageClass}`}>
-      <img src={photoURL || 'https://api.adorable.io/avatars/23/abott@adorable.png'} alt="Profile pic"/>
-      <p>{text}</p>
-    </div>
-  </>)
-};
-
 
 export default App;
